Revert "Tratamentos de error"
This reverts commit 6888387131348acf4a438976ffadbcafb790bd5f.

+++ dia-2/RaizQuadradaSimples.py
@@ -1,6 +1,6 @@
 import math;
 
-numero = int(input("Digite um número: "));
+numero = float(input("Digite um número: "));
 
 raiz = math.sqrt(numero);
